Extract awa-metrics crate (closes #176)
Carve out an internal awa-metrics crate so awa-ui and other non-runtime
callers can emit Awa's OTel counters with the exact same metric names and
attribute sets the worker uses, without depending on awa-worker (which
pulls in the dispatcher/runtime crate graph for a few counter increments).

- New publish=false awa-metrics crate. Minimal deps: opentelemetry +
  awa-model (for StorageStatus / RotateOutcome / PruneOutcome shared by
  the record_* methods).
- AwaMetrics + every record_* method moves verbatim from
  awa-worker/src/metrics.rs into awa-metrics/src/lib.rs.
- awa-worker/src/metrics.rs is now a single-line shim that
  pub use awa_metrics::AwaMetrics, so awa_worker::AwaMetrics and
  awa_worker::metrics::AwaMetrics keep working for awa-cli and
  awa-python (no semver break).
- awa-ui adds awa-metrics, drops opentelemetry, and replaces the
  inline emit_counter helper + ad-hoc bulk_move counter build in
  handlers/dlq.rs with AwaMetrics::from_global().record_dlq_*. The
  comment-only "must stay in sync with AwaMetrics" invariant is now
  enforced by the type system instead.
- awa-metrics also exposes a names module with public string
  constants for every metric, for tests asserting against an OTLP
  exporter without copy-pasting literals.

No behavior change. Validation:
- cargo build --workspace (clean)
- cargo clippy --workspace --all-targets -- -D warnings (clean)
- cargo fmt --all
- cargo test -p awa-metrics (existing record_*_info tests pass)
