+++ awa-worker/Cargo.toml
@@ -16,6 +16,7 @@ __python-bridge = []
 [dependencies]
 awa-model.workspace = true
 awa-macros.workspace = true
+awa-metrics.workspace = true
 sqlx.workspace = true
 serde.workspace = true
 serde_json.workspace = true
